+++ src/app/app.component.ts
@@ -4,6 +4,7 @@ import {StatusBar, Splashscreen} from 'ionic-native';
 import {ListsPage} from "../pages/lists/lists";
 import {TrashPage} from "../pages/trash/trash";
 import {Storage} from '@ionic/storage';
+import {IntroPage} from "../pages/intro/intro";
 
 const PAGES = [
   {name: 'Lists', component: ListsPage},
@@ -16,7 +17,7 @@ const PAGES = [
 
 export class MyApp {
   @ViewChild(Nav) nav;
-  rootPage = ListsPage;
+  rootPage;
 
   constructor(platform: Platform, public storage: Storage) {
     platform.ready().then(() => {
@@ -25,6 +26,15 @@ export class MyApp {
       //StatusBar.styleDefault();
       //Splashscreen.hide();
     });
+
+    this.storage.get('firstLaunch').then(v => {
+      if (v === null || v === undefined || v === true) {
+        this.storage.set('firstLaunch', true);
+        this.rootPage = IntroPage;
+      } else {
+        this.rootPage = ListsPage
+      }
+    })
 
     this.storage.get('lists').then(data => {
       if (!data) {