+++ src/pages/trash/trash.ts
@@ -1,5 +1,5 @@
 import {Component} from '@angular/core';
-import {NavController, NavParams, ModalController} from 'ionic-angular';
+import {NavController, NavParams, ModalController, AlertController, ToastController} from 'ionic-angular';
 import {LISTS} from "../../models/mock";
 import {Lists} from "../../providers/lists";
 
@@ -14,7 +14,9 @@ export class TrashPage {
   constructor(public navCtrl: NavController,
               public navParams: NavParams,
               public modalCtrl: ModalController,
-              public listsService: Lists) {
+              public listsService: Lists,
+              public alertCtrl: AlertController,
+              public toastCtrl: ToastController) {
     this.listsService.getLists().then(data => {
       this.lists = JSON.parse(data).filter(list => list.removed);
     });
@@ -22,6 +24,42 @@ export class TrashPage {
 
   ionViewDidLoad() {
 
+  }
+
+  deleteAll() {
+    let alert = this.alertCtrl.create({
+      title: 'Delete forever',
+      message: `Delete ${this.lists.length} lists forever?`,
+      buttons: [
+        {
+          text: 'Cancel',
+          role: 'cancel',
+        },
+        {
+          text: 'Delete',
+          handler: () => {
+            this.showToast(`${this.lists.length} lists has been removed forever`);
+            this.lists.forEach(list => {
+              this.listsService.deleteList(list);
+              console.log('lul');
+            });
+          }
+        }
+      ]
+    });
+
+    alert.present();
+  }
+
+  showToast(message) {
+    let toast = this.toastCtrl.create({
+      message: message,
+      duration: 3000,
+      position: 'bottom',
+      showCloseButton: true,
+      closeButtonText: 'OK'
+    });
+    toast.present();
   }
 
 }