+++ src/pages/view-list/view-list.ts
@@ -1,7 +1,7 @@
 import {Component} from '@angular/core';
 import {
   NavController, NavParams, ModalController, ActionSheetController, Platform,
-  AlertController
+  AlertController, ToastController
 } from 'ionic-angular';
 import {ListFormPage} from "../list-form/list-form";
 import {Lists} from "../../providers/lists";
@@ -23,7 +23,8 @@ export class ViewListPage {
               public listsService: Lists,
               public actionSheetCtrl: ActionSheetController,
               public alertCtrl: AlertController,
-              public helpers: Helpers) {
+              public helpers: Helpers,
+              public toastCtrl: ToastController) {
     this.list = this.navParams.get('list');
   }
 
@@ -33,6 +34,11 @@ export class ViewListPage {
 
   editList(list) {
     let editModal = this.modalCtrl.create(ListFormPage, {list: list, mode: 'edit', title: 'Edit List'});
+
+    editModal.onDidDismiss(() => {
+      this.showToast('List successfully updated');
+    });
+
     editModal.present();
   }
 
@@ -85,6 +91,7 @@ export class ViewListPage {
             list.removed = true;
             this.listsService.updateList(list);
             this.navCtrl.pop();
+            this.showToast(`${list.title} moved to trash`)
           }
         }
       ]
@@ -98,6 +105,7 @@ export class ViewListPage {
       if (list.color !== color) {
         list.color = color;
         this.listsService.updateList(list);
+        this.showToast('List successfully updated');
       }
     });
 
@@ -132,11 +140,23 @@ export class ViewListPage {
               createdAt: Date.now(),
             });
             this.listsService.updateList(list);
+            this.showToast('List successfully updated');
           }
         }
       ]
     });
     alert.present();
+  }
+
+  showToast(message) {
+    let toast = this.toastCtrl.create({
+      message: message,
+      duration: 3000,
+      position: 'bottom',
+      showCloseButton: true,
+      closeButtonText: 'OK'
+    });
+    toast.present();
   }
 
   ionViewDidLoad() {